requirements.txt
aiofiles
asyncio
httpx
pyyaml
